linkedin-agents/requirements.txt
chromadb==0.5.0
langchain==0.1.17
langchain_anthropic==0.1.11
langchain_community==0.0.36
langchain_core==0.1.50
langgraph==0.0.44
playwright==1.42.0
pydantic==2.7.1
python-dotenv==1.0.1
schedule==1.2.1
st_pages==0.4.5
streamlit==1.32.2
